feat: Dynamically load GraphiQL Playground
